compress CISDI_Model, Position type Float4->Float3, Normal type Float2 -> Int16_2, UV type Float2 -> UInt16_2.

+++ Source/Core/Application/Application.h
@@ -91,7 +91,7 @@ protected:
 
     
     bool mStopRendering {false};
-    uint32_t mFrameNum {0};
+    uint64_t mFrameNum {0};
     IntelliDesign_NS::Core::Utils::FrameTimer mFrameTimer;
 
 #ifdef CUDA_VULKAN_INTEROP
